Finish parsing and compilationn
