+++ json/src/main/scala/fs2/data/json/jq/ast.scala
@@ -16,10 +16,20 @@
 
 package fs2.data.json.jq
 
-import cats.data.NonEmptyList
+import cats.data.NonEmptyChain
 
 sealed trait Jq
-sealed trait Filter extends Jq
+sealed trait Filter extends Jq {
+  def ~(that: Filter): Filter =
+    (this, that) match {
+      case (Jq.Identity, _) => that
+      case (_, Jq.Identity) => this
+      case (Jq.Sequence(s1), Jq.Sequence(s2)) => Jq.Sequence(s1 ++ s2)
+      case (Jq.Sequence(s1), _) => Jq.Sequence(s1 :+ that)
+      case (_, Jq.Sequence(s2)) => Jq.Sequence(this +: s2)
+      case (_, _) => Jq.Sequence(NonEmptyChain(this, that))
+    }
+}
 sealed trait Constructor extends Jq
 object Jq {
   // filters
@@ -28,13 +38,13 @@ object Jq {
   final case class Index(idx: Int) extends Filter
   final case class Slice(start: Int, end: Option[Int]) extends Filter
   final case object RecursiveDescent extends Filter
-  final case class Sequence(jqs: NonEmptyList[Filter]) extends Filter
+  final case class Sequence(jqs: NonEmptyChain[Filter]) extends Filter
 
   final case class Iterator(filter: Filter, inner: Jq) extends Jq
 
   // constructors
-  final case class Arr(values: List[Jq]) extends Constructor
-  final case class Obj(fields: List[(String, Jq)]) extends Constructor
+  final case class Arr(prefix: Filter, values: List[Jq]) extends Constructor
+  final case class Obj(prefix: Filter, fields: List[(String, Jq)]) extends Constructor
   final case class Num(n: String) extends Constructor
   final case class Str(s: String) extends Constructor
   final case class Bool(b: Boolean) extends Constructor
